Converted aureliamof to use indexedDB instead of localstorage
Localstorage has a limit of 5MB per domain name. IndexedDB seems to be unlimited. So data is now stored in an indexedDB for fast retrieval on repeat visits, rather than in localstorage. This should clear up any problems when viewing more than one MOF-viewer on the same browser.
Code is a lot more complicated as indexedDB calls are asynchronous (localstorage calls are synchronous)